+++ sparrow-sharp/Sparrow/Utils/ColorUtil.cs
@@ -4,25 +4,29 @@ namespace Sparrow.Utils
 {
 	public class ColorUtil
 	{
-		public static uint GetRGB (uint r, uint g, uint b)
+		public static uint GetRGB (byte r, byte g, byte b)
 		{
-			return (r << 16 | g << 8 | b);
+			return (uint)(r << 16 | g << 8 | b);
 		}
 
-		public static uint GetA(uint color) {
-			return ((color >> 24) & 0xFF);
+		public static byte GetA (uint color)
+		{
+			return (byte)((color >> 24) & 0xFF);
 		}
 
-		public static uint GetR(uint color) {
-			return ((color >> 16) & 0xFF);
+		public static byte GetR (uint color)
+		{
+			return (byte)((color >> 16) & 0xFF);
 		}
 
-		public static uint GetG(uint color) {
-			return ((color >> 8) & 0xFF);
+		public static byte GetG (uint color)
+		{
+			return (byte)((color >> 8) & 0xFF);
 		}
 
-		public static uint GetB(uint color) {
-			return (color & 0xFF);
+		public static byte GetB (uint color)
+		{
+			return (byte)(color & 0xFF);
 		}
 	}
 }
