+++ sparrow-sharp/Sparrow/Utils/NumberUtil.cs
@@ -4,17 +4,17 @@ namespace Sparrow.Utils
 {
 	public class NumberUtil
 	{
-		public static bool Equals (double a, double b)
+		public static bool Equals (float a, float b)
 		{
-			return Equals (a, b, 0.000001);
+			return Equals (a, b, 0.000001f);
 		}
 
-		public static bool Equals (double a, double b, double tolerance)
+		public static bool Equals (float a, float b, float tolerance)
 		{
 			return Math.Abs (a - b) < tolerance;
 		}
 
-		public static double Clamp(double value, double min, double max) {
+		public static float Clamp(float value, float min, float max) {
 			return Math.Min (max, Math.Max (value, min));
 		}
 	}
